Refactor. Move GH issues util to a different file
